♻️ refactor(fire_control): 简化火控架构
1. 移除冗余控制器
   - 删除 PidController (未使用)
   - 删除 MpcController (功能合并到 GimbalPlanner)
   - 删除 FireStrategy 接口

2. 新增 FireDecision 模块
   - 独立的开火决策逻辑
   - 支持弹道误差阈值判断

3. 重构 FireController
   - 简化为单一职责: 目标选择 + 弹道解算 + 开火决策
   - 集成 GimbalPlanner 进行云台规划
   - 改进延迟补偿和位置预测

4. 简化 GimbalPlanner
   - 移除冗余的 MPC 状态管理
   - 专注于轨迹优化核心功能

5. 添加火控单元测试

+++ aimer/auto_aim/fire_control/fire_controller_node.cpp
@@ -102,7 +102,7 @@ void fire_control_run(const std::string& /* config_path */)
                 distance = armor->position.norm();
             }
         }
-        double bullet_speed = std::max(snapshot.self_state.bullet_speed, 10.0);
+        double bullet_speed = std::max(static_cast<double>(snapshot.self_state.bullet_speed), 10.0);
         latency.fire_to_hit = distance / bullet_speed;
 
         // ========== 执行控制 ==========
